Reworked event capture mechanism
Previous testing were not as robust, so we lost a couple of edge cases.
Most "send characters to user" functions are refactored to
`character_sender`. This file is coupled with `keyboard_controller`.
We also intercepted modifier keys early in `keyboard_hook`, since the
`GetKeyboardState` and its variants are not reliable when called from
hooks (they work normally on message queue). This requires us to edit
the keystate manually before sending it to `ToUnicodeEx`.